--- a/cache/clock_cache.cc
+++ b/cache/clock_cache.cc
@@ -272,15 +272,13 @@ class ClockCacheShard final : public CacheShard {
                 void (*deleter)(const Slice& key, void* value),
                 Cache::Handle** handle, Cache::Priority priority) override;
   Status Insert(const Slice& key, uint32_t hash, void* value,
-                Cache::CacheItemHelperCallback helper_cb, size_t charge,
+                const Cache::CacheItemHelper* helper, size_t charge,
                 Cache::Handle** handle, Cache::Priority priority) override {
-    Cache::DeletionCallback delete_cb;
-    (*helper_cb)(nullptr, nullptr, &delete_cb);
-    return Insert(key, hash, value, charge, delete_cb, handle, priority);
+    return Insert(key, hash, value, charge, helper->del_cb, handle, priority);
   }
   Cache::Handle* Lookup(const Slice& key, uint32_t hash) override;
   Cache::Handle* Lookup(const Slice& key, uint32_t hash,
-                        Cache::CacheItemHelperCallback /*helper_cb*/,
+                        const Cache::CacheItemHelper* /*helper*/,
                         const Cache::CreateCallback& /*create_cb*/,
                         Cache::Priority /*priority*/, bool /*wait*/) override {
     return Lookup(key, hash);