… cache/lru_cache_test.cc
… cache/sharded_cache.h
… db/db_basic_test.cc
… db/db_block_cache_test.cc
… db/db_test_util.h
… include/rocksdb/cache.h
… include/rocksdb/secondary_cache.h
… include/rocksdb/tiered_cache.h
… utilities/simulator_cache/sim_cache.cc