@pdillinger 2021-05-06
> It seems like this would involve exposing some CacheableEntry / BlocklikeTraits functionality into block cache (seems reasonable) ...

Except perhaps for Cache API compatibility, which I would argue is not so critical.